requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
langchain-openai
streamlit
